Connect Dependencies with ObjPooling and UI
