+++ Assets/Prefabs/Dependencies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e4b03706b7060c94d855e35a5da5fa14
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
