+++ Assets/Prefabs/Dependencies/Dependencies.prefab
@@ -1,6 +1,6 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
---- !u!1 &8113515734996193455
+--- !u!1 &3696146047767597133
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -8,42 +8,39 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 8113515734996193453}
-  - component: {fileID: 8113515734996193452}
+  - component: {fileID: 7253993208163888508}
+  - component: {fileID: 6800250808489893674}
   m_Layer: 0
-  m_Name: Pool
+  m_Name: Dependencies
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!4 &8113515734996193453
+--- !u!4 &7253993208163888508
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8113515734996193455}
+  m_GameObject: {fileID: 3696146047767597133}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalPosition: {x: 363.12857, y: 23.123037, z: -5.5363097}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!114 &8113515734996193452
+--- !u!114 &6800250808489893674
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 8113515734996193455}
+  m_GameObject: {fileID: 3696146047767597133}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: f160762e9f1b67046988d9628e3d9e5b, type: 3}
+  m_Script: {fileID: 11500000, guid: ede6413906e01d649b07f383a5f5bf48, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  amount: 10
-  pool: {fileID: 11400000, guid: f2fa46f2fd8731247aa1dd9f20fea179, type: 2}
-  enemyHand: {fileID: 11400000, guid: 14116d48ebbdcfa468e7417857de4e33, type: 2}
