+++ Assets/Prefabs/Dependencies/Dependencies.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0457222e79fb9654b865f8c1bb697fdf
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
